Cleaning base model class. Moving Parse items to a LoadableModel class. Making most functions pure virtual. Cleanups to all models classes considering the new pure virtual functions. SMall Cleanups throughout.
git-svn-id: svn+ssh://parasol-svn.cs.tamu.edu/research/parasol-svn/svnrepository/vizmo/trunk@479 8e255c04-9930-0410-9bf1-9572f6c5958d

+++ src/Utilities/GL/GLUtilities.h
@@ -10,7 +10,7 @@
 inline void
 DrawRotateAxis(Camera* _camera) {
   static GLuint gid = -1;
-  if(gid == -1) { //no gid is created
+  if(gid == (GLuint)-1) { //no gid is created
     gid = glGenLists(1);
     glNewList(gid, GL_COMPILE);
